Extract module-level tests from test_metrics.py
Move stale PR tests to test__stale_pr.py and AI detection tests
to test__ai_detection.py. test_metrics.py drops from 939 to 689 lines.
Fix missing assertion in test_should_count_pr_with_trailer_only_in_commit_message.
