+++ tests/unit/metrics/test_metrics.py
@@ -1,7 +1,5 @@
 """Unit tests for reports.py functions."""
 
-from typing import cast
-
 from git_dev_metrics.metrics import (
     calculate_avg_lines_per_pr,
     calculate_cycle_time,
@@ -12,7 +10,6 @@ from git_dev_metrics.metrics import (
     calculate_throughput,
     median,
 )
-from git_dev_metrics.models import OpenPullRequest
 from git_dev_metrics.utils import TimePeriod
 from git_dev_metrics.utils import period_days as _period_days
 
@@ -684,255 +681,3 @@ class TestGroupPrsByDevsBotExclusion:
         result = group_prs_by_devs(prs)
         assert "alice" in result
         assert "patches-bot" not in result
-
-
-class TestGetStalePrs:
-    def test_should_return_empty_for_empty_list(self):
-        from git_dev_metrics.metrics._stale_pr import get_stale_prs
-
-        result = get_stale_prs([], "myrepo")
-        assert result == []
-
-    def test_should_return_fresh_prs(self):
-        from datetime import UTC, datetime, timedelta
-
-        from git_dev_metrics.metrics._stale_pr import get_stale_prs
-
-        now = datetime.now(UTC)
-        prs = cast(
-            list[OpenPullRequest],
-            [
-                {
-                    "number": 1,
-                    "title": "Fresh PR",
-                    "created_at": now - timedelta(days=1),
-                    "merged_at": None,
-                    "user": {"login": "alice"},
-                },
-            ],
-        )
-        result = get_stale_prs(prs, "myrepo", lambda: now)
-        assert result == []
-
-    def test_should_identify_stale_prs(self):
-        from datetime import UTC, datetime, timedelta
-
-        from git_dev_metrics.metrics._stale_pr import get_stale_prs
-
-        now = datetime.now(UTC)
-        prs = cast(
-            list[OpenPullRequest],
-            [
-                {
-                    "number": 1,
-                    "title": "Stale PR",
-                    "created_at": now - timedelta(days=10),
-                    "merged_at": None,
-                    "user": {"login": "alice"},
-                },
-            ],
-        )
-        result = get_stale_prs(prs, "myrepo", lambda: now)
-        assert len(result) == 1
-        assert result[0].number == 1
-        assert result[0].author == "alice"
-        assert result[0].repo == "myrepo"
-        assert result[0].age_hours > 24 * 7
-
-    def test_should_sort_by_age_oldest_first(self):
-        from datetime import UTC, datetime, timedelta
-
-        from git_dev_metrics.metrics._stale_pr import get_stale_prs
-
-        now = datetime.now(UTC)
-        prs = cast(
-            list[OpenPullRequest],
-            [
-                {
-                    "number": 1,
-                    "title": "Newer stale",
-                    "created_at": now - timedelta(days=8),
-                    "merged_at": None,
-                    "user": {"login": "alice"},
-                },
-                {
-                    "number": 2,
-                    "title": "Older stale",
-                    "created_at": now - timedelta(days=15),
-                    "merged_at": None,
-                    "user": {"login": "bob"},
-                },
-            ],
-        )
-        result = get_stale_prs(prs, "myrepo", lambda: now)
-        assert result[0].number == 2
-        assert result[1].number == 1
-        assert result[0].repo == "myrepo"
-
-
-class TestSummarizeStalePrs:
-    def test_should_return_zero_for_empty_list(self):
-        from git_dev_metrics.metrics._stale_pr import summarize_stale_prs
-
-        assert summarize_stale_prs([]) == (0, 0.0)
-
-    def test_should_compute_count_and_mean_age(self):
-        from git_dev_metrics.metrics._stale_pr import StalePr, summarize_stale_prs
-
-        prs = [
-            StalePr(
-                number=1,
-                title="a",
-                author="x",
-                repo="r",
-                age_hours=480,
-                age_days=20.0,
-                is_draft=False,
-                is_approved=False,
-                url="",
-            ),
-            StalePr(
-                number=2,
-                title="b",
-                author="y",
-                repo="r",
-                age_hours=240,
-                age_days=10.0,
-                is_draft=False,
-                is_approved=False,
-                url="",
-            ),
-        ]
-        assert summarize_stale_prs(prs) == (2, 15.0)
-
-
-class TestIsAiCoauthored:
-    """Test cases for is_ai_coauthored function."""
-
-    def test_should_return_false_for_none_body(self):
-        from git_dev_metrics.metrics._ai_detection import is_ai_coauthored
-
-        result = is_ai_coauthored(any_pr(body=None))
-        assert result is False
-
-    def test_should_return_false_for_empty_body(self):
-        from git_dev_metrics.metrics._ai_detection import is_ai_coauthored
-
-        result = is_ai_coauthored(any_pr(body=""))
-        assert result is False
-
-    def test_should_return_false_for_no_trailer(self):
-        from git_dev_metrics.metrics._ai_detection import is_ai_coauthored
-
-        result = is_ai_coauthored(any_pr(body="This is a regular PR description"))
-        assert result is False
-
-    def test_should_return_true_for_coauthored_by_in_body(self):
-        from git_dev_metrics.metrics._ai_detection import is_ai_coauthored
-
-        result = is_ai_coauthored(any_pr(body="Co-Authored-By: GitHub <noreply@github.com>"))
-        assert result is True
-
-    def test_should_be_case_insensitive(self):
-        from git_dev_metrics.metrics._ai_detection import is_ai_coauthored
-
-        result = is_ai_coauthored(any_pr(body="co-authored-by: someone@example.com"))
-        assert result is True
-
-    def test_should_return_true_when_trailer_only_in_commit_message(self):
-        from git_dev_metrics.metrics._ai_detection import is_ai_coauthored
-
-        pr = any_pr(
-            body=None,
-            commit_messages=["Fix bug\n\nCo-Authored-By: Claude <noreply@anthropic.com>"],
-        )
-        result = is_ai_coauthored(pr)
-        assert result is True
-
-    def test_should_return_true_when_trailer_in_one_of_many_commits(self):
-        from git_dev_metrics.metrics._ai_detection import is_ai_coauthored
-
-        pr = any_pr(
-            body="Regular PR",
-            commit_messages=[
-                "Refactor module",
-                "Add tests",
-                "Polish\n\n🤖 Generated with Claude Code",
-            ],
-        )
-        result = is_ai_coauthored(pr)
-        assert result is True
-
-    def test_should_return_false_when_no_trailer_in_body_or_commits(self):
-        from git_dev_metrics.metrics._ai_detection import is_ai_coauthored
-
-        pr = any_pr(
-            body="Regular PR",
-            commit_messages=["Refactor module", "Add tests"],
-        )
-        result = is_ai_coauthored(pr)
-        assert result is False
-
-
-class TestCalculateAiPercentage:
-    """Test cases for calculate_ai_percentage function."""
-
-    def test_should_return_zero_for_empty_list(self):
-        from git_dev_metrics.metrics._ai_detection import calculate_ai_percentage
-
-        result = calculate_ai_percentage([])
-        assert result == 0.0
-
-    def test_should_return_zero_for_no_ai_prs(self):
-        from git_dev_metrics.metrics._ai_detection import calculate_ai_percentage
-
-        prs = [
-            any_pr(body=None),
-            any_pr(body="Regular PR"),
-            any_pr(body=None),
-        ]
-        result = calculate_ai_percentage(prs)
-        assert result == 0.0
-
-    def test_should_calculate_percentage_correctly(self):
-        from git_dev_metrics.metrics._ai_detection import calculate_ai_percentage
-
-        prs = [
-            any_pr(body="Co-Authored-By: someone@example.com"),
-            any_pr(body="Regular PR"),
-            any_pr(body="Co-Authored-By: another@example.com"),
-            any_pr(body="Another regular PR"),
-        ]
-        result = calculate_ai_percentage(prs)
-        assert result == 50.0
-
-    def test_should_handle_all_ai_prs(self):
-        from git_dev_metrics.metrics._ai_detection import calculate_ai_percentage
-
-        prs = [
-            any_pr(body="Co-Authored-By: someone@example.com"),
-            any_pr(body="Co-Authored-By: another@example.com"),
-        ]
-        result = calculate_ai_percentage(prs)
-        assert result == 100.0
-
-    def test_should_round_to_one_decimal(self):
-        from git_dev_metrics.metrics._ai_detection import calculate_ai_percentage
-
-        prs = [
-            any_pr(body="Co-Authored-By: someone@example.com"),
-            any_pr(body="Regular PR"),
-            any_pr(body="Regular PR"),
-        ]
-        result = calculate_ai_percentage(prs)
-        assert result == 33.3
-
-    def test_should_count_pr_with_trailer_only_in_commit_message(self):
-        from git_dev_metrics.metrics._ai_detection import calculate_ai_percentage
-
-        prs = [
-            any_pr(body="Regular PR", commit_messages=["Co-Authored-By: Claude"]),
-            any_pr(body="Regular PR", commit_messages=["Refactor"]),
-        ]
-        result = calculate_ai_percentage(prs)
-        assert result == 50.0
